+++ appseed/aura/primitive/datetime/datetime_c.h
@@ -5,7 +5,7 @@ BEGIN_EXTERN_C
 
 
 CLASS_DECL_AURA DWORD get_first_tick();
-CLASS_DECL_AURA DWORD get_first_nano();
+CLASS_DECL_AURA uint64_t get_first_nano();
 CLASS_DECL_AURA DWORD get_tick_count();
 CLASS_DECL_AURA uint64_t get_nanos();
